chore(interopgen): move depset package out of op-supervisor
Relocates op-supervisor/supervisor/backend/depset to
op-chain-ops/interopgen/depset. The package is a set of config value
types (FullConfigSet, StaticRollupConfigSet, etc.) consumed by
op-chain-ops/interopgen, op-devstack/sysgo, and op-program — none of
which are op-supervisor consumers. Moving it removes one of the
remaining reasons op-supervisor must exist as a Go module dependency.
